MGS-4515 [#ccc] integrate 6.4.0.alpha release package
This is the first 6.4.0 release for code preview.
Test coverage
  1)	Boards: IMX8
  2)	OS: Android 9.0, wayland
  3)	QnX need sanity test
Testing:
  1)	Conformance (OVX1.2/Vulkan1.1/GLES3.2.5/OCL, may left failure, but no hang)
  2)	VSI focusing on VK1.1 new feature, thus Vulkan will certainly have issue on both 1.0.2 and 1.1
	- Not only failures, also found Vulkan hang on Wayland
  3)	Verify OpenCL/VX dual GPU on 8QM
	- OpenVX1.2 CTS can pass on Wayland
	- OpenCL CTS has regression, expecting fix
  4)	Android NN on 8QM
	- Pass a few tests, met freeze/hang, checking with Android NN CTS
  5)	Android NN for QXP/mScale ?
	- VSI will support Android NN API on QXP/mScale GPU SH in 64x driver, Feb, 2019
  6)	IMX-1275  Acceptance list for Vision test for IMX8 boards
	- Wayland test is same with 624, VSI is looking NN related part,
  7)	VIPNanoSi– unified driver with GPU(HW will be not ready, VSI will use either VIP ready HW to test)
	- Verify OpenCL/OpenVX CTS, Vivante NN tests
	- Reference board is not available, will run FPGA(CTS first)
	- Two failures with OpenVX 1.2 cts, OpenCL is still running

+++ driver/compiler/libVSC/debug/gc_vsc_debug.c
@@ -1,6 +1,6 @@
 /****************************************************************************
 *
-*    Copyright (c) 2005 - 2018 by Vivante Corp.  All rights reserved.
+*    Copyright (c) 2005 - 2019 by Vivante Corp.  All rights reserved.
 *
 *    The material in this file is confidential and contains trade secrets
 *    of Vivante Corporation. This is proprietary information owned by
@@ -1295,7 +1295,7 @@ void vscDIStep(void * ptr, unsigned int pc, unsigned int stepFlag)
 {
     VSC_DIContext * context = (VSC_DIContext *)ptr;
 
-    if (context == gcvNULL || pc < 0)
+    if (context == gcvNULL || (int)pc < 0)
         return;
 
     context->stepState = (VSC_STEP_STATE)stepFlag;
